Socket reuse options
If bert_rpc_exec is started with the exo option

{reuse_mode, client | server}

Then the client and server will allow for reuse of the same
socket session for all RPCs in both directions.

The client API module nice_bert_rpc uses an existing session,
if available. Otherwise, the rpc_exec socket_rpc_server will
create a new one, calling exo_socket:connect(...) with the same
parameters as the client process would have. It then returns the
pid to the client, which can call the exo_socket_session process
directly with a gen_server:call().

The server, in reuse mode, doesn't allow rpc requests to the
client side, unless a connection is already established. In other
words, the client side must be the one setting up the connection.